:truck: Moved activities
