+++ base/src/main/res/values/strings.xml
